Refactor `qmk compile-json` to `qmk compile` (#6592)
